bumping version number
Because we have the new methods that kind of replace most of the
usefulness of the module.

I think they should really be in their own module since they don't
LazyInit, but I haven't thought of a name for it yet.

I also fixed a line that broke everything.. so that's important :)